+++ homeassistant/components/mill/climate.py
@@ -14,7 +14,7 @@ from homeassistant.components.climate import (
     HVACMode,
 )
 from homeassistant.const import ATTR_TEMPERATURE, PRECISION_TENTHS, UnitOfTemperature
-from homeassistant.core import HomeAssistant, ServiceCall, callback
+from homeassistant.core import HomeAssistant, callback
 from homeassistant.helpers import config_validation as cv
 from homeassistant.helpers.device_registry import CONNECTION_NETWORK_MAC, DeviceInfo
 from homeassistant.helpers.entity_platform import AddConfigEntryEntitiesCallback
@@ -26,12 +26,10 @@ from .const import (
     ATTR_ROOM_NAME,
     ATTR_SLEEP_TEMP,
     CONNECTION_TYPE,
-    DOMAIN,
     LOCAL,
     MANUFACTURER,
     MAX_TEMP,
     MIN_TEMP,
-    SERVICE_SET_ROOM_TEMP,
 )
 from .coordinator import MillConfigEntry, MillDataUpdateCoordinator
 from .entity import MillBaseEntity
@@ -64,21 +62,6 @@ async def async_setup_entry(
         if isinstance(mill_device, mill.Heater)
     ]
     async_add_entities(entities)
-
-    async def set_room_temp(service: ServiceCall) -> None:
-        """Set room temp."""
-        room_name = service.data.get(ATTR_ROOM_NAME)
-        sleep_temp = service.data.get(ATTR_SLEEP_TEMP)
-        comfort_temp = service.data.get(ATTR_COMFORT_TEMP)
-        away_temp = service.data.get(ATTR_AWAY_TEMP)
-        await mill_data_coordinator.mill_data_connection.set_room_temperatures_by_name(
-            room_name, sleep_temp, comfort_temp, away_temp
-        )
-
-    # pylint: disable-next=home-assistant-service-registered-in-setup-entry
-    hass.services.async_register(
-        DOMAIN, SERVICE_SET_ROOM_TEMP, set_room_temp, schema=SET_ROOM_TEMP_SCHEMA
-    )
 
 
 class MillHeater(MillBaseEntity, ClimateEntity):
